Move upstream prices handler out of admin
