+++ internal/handler/admin.go
@@ -5,7 +5,6 @@ import (
 	"encoding/json"
 	"errors"
 	"fmt"
-	"io"
 	"log"
 	"net/http"
 	"net/url"
@@ -18,7 +17,6 @@ import (
 	maxxctx "github.com/awsl-project/maxx/internal/context"
 	"github.com/awsl-project/maxx/internal/cooldown"
 	"github.com/awsl-project/maxx/internal/domain"
-	"github.com/awsl-project/maxx/internal/modelpriceupstream"
 	"github.com/awsl-project/maxx/internal/pricing"
 	"github.com/awsl-project/maxx/internal/repository"
 	"github.com/awsl-project/maxx/internal/service"
@@ -2526,43 +2524,6 @@ func (h *AdminHandler) handleModelPricesReset(w http.ResponseWriter, r *http.Req
 	// Refresh calculator cache
 	pricing.GlobalCalculator().LoadFromDatabase(prices)
 	writeJSON(w, http.StatusOK, prices)
-}
-
-// handleModelPricesUpstreamPrices handles POST /admin/model-prices/upstream/prices
-func (h *AdminHandler) handleModelPricesUpstreamPrices(w http.ResponseWriter, r *http.Request) {
-	req, ok := decodeModelPriceUpstreamRequest(w, r)
-	if !ok {
-		return
-	}
-	result, err := h.svc.ListModelPricesFromExternalSource(r.Context(), req.Source)
-	if err != nil {
-		status := http.StatusInternalServerError
-		if errors.Is(err, modelpriceupstream.ErrUnsupportedSource) {
-			status = http.StatusBadRequest
-		}
-		writeJSON(w, status, map[string]string{"error": err.Error()})
-		return
-	}
-	writeJSON(w, http.StatusOK, result)
-}
-
-type modelPriceUpstreamRequest struct {
-	Source string `json:"source"`
-}
-
-func decodeModelPriceUpstreamRequest(w http.ResponseWriter, r *http.Request) (modelPriceUpstreamRequest, bool) {
-	var req modelPriceUpstreamRequest
-	if r.Body == nil || r.ContentLength == 0 {
-		return req, true
-	}
-	if err := json.NewDecoder(r.Body).Decode(&req); err != nil {
-		if errors.Is(err, io.EOF) {
-			return req, true
-		}
-		writeJSON(w, http.StatusBadRequest, map[string]string{"error": "invalid request body"})
-		return req, false
-	}
-	return req, true
 }
 
 // mustGetPrices is a helper to get prices for refreshing calculator
